+++ CHANGELOG.md
@@ -358,3 +358,9 @@
 
 ## Added:
 - Added a new method `is_enemy_or_empty` to the `Piece` class. This method checks if a given position on the board is either occupied by an enemy piece or empty, improving movement validation for chess pieces. Included in `Pawn` and `King` to fix an issue of similar code. (maybe i can implement `is_enemy_or_empty` in other pieces)
+
+##[2024-10-21]
+
+## Final
+- Improve cli for the final commit
+- Little improve board and chess for final commit